feat: Foursquareの処理をFoursquareClientに移動
